+++ sertop/sertop_arg.mli
@@ -46,7 +46,9 @@ val disallow_sprop  : bool Term.t
 val indices_matter  : bool Term.t
 
 (* sertop options *)
-type comp_mode = | C_parse | C_stats | C_print | C_sexp | C_check | C_vo | C_env
+type comp_mode =
+  | C_parse | C_stats | C_print | C_sexp | C_check | C_vo | C_env | C_ktrace
+
 val comp_mode : comp_mode Term.t
 
 type comp_input = | I_vernac | I_sexp